fixed and sort algorithm

+++ app/src/main/java/com/menasy/heskit/NotWorksDay.java
@@ -4,7 +4,7 @@ public class NotWorksDay {
     private String date;
     private String reason;
     private int days;
-    private int id;
+    private long id;
 
     public NotWorksDay(int days, String date, String reason) {
         this.days = days;
@@ -15,6 +15,6 @@ public class NotWorksDay {
     public int getDays() { return days; }
     public String getDate() { return date; }
     public String getReason() { return reason != null ? reason : ""; }
-    public int getId() { return id; }
-    public void setId(int id) { this.id = id; }
+    public long getId() { return id; }
+    public void setId(long id) { this.id = id; }
 }
